feat: copy code from practice
