+++ src/fixtures/mock.fixture.ts
@@ -1,47 +1,5 @@
 import { test as base } from './controllers.fixture';
-import { Page } from '@playwright/test';
-import { apiConfig } from 'config';
-import { STATUS_CODES } from 'data';
-import { ICustomersResponse, ICustomerResponse, IMetricsResponse } from 'types';
-
-class Mock {
-  constructor(private page: Page) {}
-
-  async customers(body: ICustomersResponse, statusCode: STATUS_CODES = STATUS_CODES.OK) {
-    this.page.route(/\/api\/customers(\?.*)?$/, async (route) => {
-      await route.fulfill({
-        status: statusCode,
-        contentType: 'application/json',
-        body: JSON.stringify(body),
-      });
-    });
-  }
-
-  async customerDetails(body: ICustomerResponse, statusCode: STATUS_CODES = STATUS_CODES.OK) {
-    this.page.route(apiConfig.BASE_URL + '/' + apiConfig.ENDPOINTS.CUSTOMER_BY_ID(body.Customer._id), async (route) => {
-      await route.fulfill({
-        status: statusCode,
-        contentType: 'application/json',
-        body: JSON.stringify(body),
-      });
-    });
-  }
-
-  async metrics(body: IMetricsResponse, statusCode: STATUS_CODES = STATUS_CODES.OK) {
-    this.page.route(apiConfig.BASE_URL + '/' + apiConfig.ENDPOINTS.METRICS, async (route) => {
-      await route.fulfill({
-        status: statusCode,
-        contentType: 'application/json',
-        body: JSON.stringify(body),
-      });
-    });
-  }
-}
-
-export interface ISortingMockOptions {
-  sortField: string;
-  sortDir: string;
-}
+import { Mock } from './mock';
 
 interface MockFixture {
   mock: Mock;
