+++ src/ui/pages/modals/index.ts
@@ -1,2 +1,3 @@
 export * from './filter.modal';
 export * from './customers/deleteCustomer.modal';
+export * from './products/productDetails.modal';
