+++ src/ui/pages/modals/products/productDetails.modal.ts
@@ -1,0 +1,22 @@
+import { MANUFACTURERS } from 'data/products';
+import { IProduct } from 'types';
+import { Modal } from '../modal.page';
+
+export class ProductDetailsModal extends Modal {
+  readonly modalBody = this.page.locator(`#details-modal-body-container`);
+  readonly values = this.modalBody.locator('p');
+
+  uniqueElement = this.modalBody;
+
+  async getDetails(): Promise<Required<IProduct> & { createdOn: string }> {
+    const [name, amount, price, manufacturer, createdOn, notes] = await this.values.allInnerTexts();
+    return {
+      name,
+      amount: +amount,
+      price: +price,
+      manufacturer: manufacturer as MANUFACTURERS,
+      createdOn,
+      notes,
+    };
+  }
+}
